Hpp and Cpp files for publisher ready
