+++ include/ROS2_Beginner_Tutorials/publisher_func.hpp
@@ -1,0 +1,42 @@
+/**
+ * @file publisher_func.hpp
+ * @author Vyshnav Achuthan (vyachu07@umd.edu)
+ * @brief 
+ * @version 0.1
+ * @date 2023-11-07
+ * 
+ * @copyright Copyright (c) 2023
+ * 
+ */
+#pragma once 
+#include <chrono>
+#include <functional>
+#include <memory>
+#include <string>
+
+#include "rclcpp/rclcpp.hpp"
+#include "std_msgs/msg/string.hpp"
+
+using namespace std::chrono_literals;
+
+/**
+ * @brief Class for Publisher 
+ * 
+ */
+class MinimalPublisher : public rclcpp::Node
+{
+  public:
+    MinimalPublisher()
+    : Node("minimal_publisher"), count_(0)
+    {
+      publisher_ = this->create_publisher<std_msgs::msg::String>("topic", 10);
+      timer_ = this->create_wall_timer(
+      500ms, std::bind(&MinimalPublisher::timer_callback, this));
+    }
+  private:
+  void timer_callback();
+    rclcpp::TimerBase::SharedPtr timer_;
+    rclcpp::Publisher<std_msgs::msg::String>::SharedPtr publisher_;
+    size_t count_;
+};
+
