+++ src/publisher_func.cpp
@@ -8,40 +8,12 @@
  * @copyright Copyright (c) 2023
  * 
  */
-#include <chrono>
-#include <functional>
-#include <memory>
-#include <string>
+#include "../include/ROS2_Beginner_Tutorials/publisher_func.hpp"
 
-#include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/string.hpp"
-
-using namespace std::chrono_literals;
-
-/**
- * @brief Class for Publisher 
- * 
- */
-class MinimalPublisher : public rclcpp::Node
-{
-  public:
-    MinimalPublisher()
-    : Node("minimal_publisher"), count_(0)
-    {
-      publisher_ = this->create_publisher<std_msgs::msg::String>("topic", 10);
-      timer_ = this->create_wall_timer(
-      500ms, std::bind(&MinimalPublisher::timer_callback, this));
-    }
-
-  private:
-    void timer_callback()
+void MinimalPublisher::timer_callback()
     {
       auto message = std_msgs::msg::String();
       message.data = "Hello, world! " + std::to_string(count_++);
       RCLCPP_INFO(this->get_logger(), "Publishing: '%s'", message.data.c_str());
       publisher_->publish(message);
     }
-    rclcpp::TimerBase::SharedPtr timer_;
-    rclcpp::Publisher<std_msgs::msg::String>::SharedPtr publisher_;
-    size_t count_;
-};
